% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/install_shapefile.R
\name{install_shapefile}
\alias{install_shapefile}
\title{Translates a compressed shapefile into the glyph-ready .rda file}
\usage{
install_shapefile(name, location, zfile, id_variable = "GMI_ADMIN")
}
\description{
See \code{install_shapefiles_from_listing()} for an example of use. In addition to
creating data tables "[name]_shapefile" and "[name]_datafile" in the global environment,
this function creates a "[name].rda" file in the current working directory.
}

