% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_shapefiles.R
\name{list_shapefiles}
\alias{list_shapefiles}
\title{List the available resources}
\usage{
list_shapefiles(type = NULL)
}
\value{
a data table. Each shapefile is one case.  The shape data are available under
the \code{code} variable.
}
\description{
List the available resources
}
\seealso{
\code{get_shapefile()} \code{get_datafile()}
}

